@@ -2,29 +2,10 @@
 
 namespace Ymigval\LaravelIndexnow\Exceptions;
 
-use Exception;
-use Ymigval\LaravelIndexnow\LogManager;
-
-class MixedException extends Exception
+class MixedException extends BaseLoggingException
 {
-    /**
-     * @var string
-     */
-    protected $message;
-
-    /**
-     * @var int
-     */
-    protected $code;
-
     public function __construct(string $message, int $code = 0)
     {
-        $this->message = $message;
-        $this->code = $code;
-    }
-
-    public function __destruct()
-    {
-        LogManager::addLog($this->message);
+        parent::__construct($message, $code);
     }
 }
