@@ -5,7 +5,6 @@ namespace Ymigval\LaravelIndexnow;
 use Exception;
 use Illuminate\Support\Facades\App;
 use Illuminate\Support\Facades\Config;
-use Illuminate\Support\Facades\File;
 use Illuminate\Support\Facades\Http;
 use Illuminate\Support\Str;
 use Ymigval\LaravelIndexnow\Exceptions\ExcessUrlsException;
@@ -16,38 +15,32 @@ use Ymigval\LaravelIndexnow\Exceptions\SearchEngineUnknownException;
 
 class IndexNowService
 {
-    /**
-     * @var string | null
-     */
-    private string $searchengine;
-
+    private ?string $searchengine = null;
     private string $key;
-
     private ?string $keyFile = null;
-
     private array $urls = [];
 
     /**
-     * Initialize the IndexNow service.
-     *
-     * @param  string  $searchengine The Search Engine to be used for indexing.
+     * Initialize the IndexNow service with the given search engine.
      *
      * @throws SearchEngineUnknownException
      */
     public function __construct(string $searchengine)
     {
         $this->setSearchEngine($searchengine);
-        $this->key = IndexNowApiKeyManager::getApiKey();
+        $this->key = IndexNowApiKeyManager::fetchOrGenerate();
     }
 
     /**
-     * Set the Search Engine for the IndexNow service.
+     * Sets the search engine driver URL based on the provided search engine name.
      *
-     * @throws SearchEngineUnknownException
+     * @param string $searchengine The name of the search engine to be set.
+     * @return void
+     * @throws SearchEngineUnknownException If the provided search engine name is not recognized or invalid.
      */
     public function setSearchEngine(string $searchengine): void
     {
-        $this->searchengine = SearchEngine::getUrl($searchengine);
+        $this->searchengine = SearchEngine::getDriverUrl($searchengine);
 
         if (is_null($this->searchengine)) {
             throw new SearchEngineUnknownException();
@@ -55,15 +48,19 @@ class IndexNowService
     }
 
     /**
-     * Get the established searchengine for the IndexNow service.
+     * Retrieves the currently set search engine driver URL.
+     *
+     * @return string The driver URL of the current search engine or an empty string if no search engine is set.
      */
     public function getSearchEngine(): string
     {
-        return (string) $this->searchengine;
+        return $this->searchengine ?? '';
     }
 
     /**
-     * Get the IndexNow API key.
+     * Retrieves the key associated with the current instance.
+     *
+     * @return string The key stored within the instance.
      */
     public function getKey(): string
     {
@@ -71,17 +68,16 @@ class IndexNowService
     }
 
     /**
-     * Hosting a text key file within your host
+     * Sets the API key file path and validates the API key retrieved from the file.
      *
-     * @throws InvalidKeyException
+     * @param string $file The file path containing the API key.
+     * @return void
+     * @throws InvalidKeyException If the API key is invalid, has incorrect length, or contains line breaks.
      */
     public function keyFile(string $file): void
     {
         $apiKey = Http::get($file)->body();
-        $apiKeyLength = strlen($apiKey);
-        $countNewLines = substr_count($apiKey, "\n");
-
-        if ($countNewLines != 0 || $apiKeyLength < 8 || $apiKeyLength > 128) {
+        if (substr_count($apiKey, "\n") !== 0 || strlen($apiKey) < 8 || strlen($apiKey) > 128) {
             throw new InvalidKeyException();
         }
 
@@ -90,19 +86,21 @@ class IndexNowService
     }
 
     /**
-     * Set URL for indexing.
+     * Adds the provided URL to the list of URLs and returns the current instance.
      *
-     * @return $this
+     * @param string $url The URL to be added.
+     * @return static Returns the current instance for method chaining.
      */
-    public function setUrl(string $url): IndexNowService
+    public function setUrl(string $url): static
     {
         $this->urls[] = $url;
-
         return $this;
     }
 
     /**
-     * Get URLs to be indexed.
+     * Retrieves a unique list of URLs stored in the instance.
+     *
+     * @return array An array containing the unique URLs.
      */
     public function getUrls(): array
     {
@@ -110,137 +108,141 @@ class IndexNowService
     }
 
     /**
-     * Submit one or more URLs for indexing.
+     * Processes the submission for the provided URLs and returns the result.
      *
-     * @param  string|array|null  $url
-     * @return string|array
+     * @param array|string|null $url The URL(s) to be processed. Can be a string, an array of strings, or null.
+     * @return array|string The result of the processing, either as an array or a string.
+     * @throws MixedException
      */
-    public function submit($url = null)
+    public function submit(array|string $url = null): array|string
     {
-        if (! is_null($url)) {
+        if ($url !== null) {
             $this->setUrls($url);
         }
 
         $process = $this->process();
-
-        LogManager::addLog($process);
+        LogManager::addMessage($process);
 
         return $process;
     }
 
-    /**
-     * Set URLs for indexing.
-     *
-     * @param  string|array  $url
-     */
-    private function setUrls($url): void
+    private function setUrls(array|string $url): void
     {
-        if (is_array($url)) {
-            $this->urls = array_merge($this->urls, $url);
-        } else {
-            $this->urls[] = $url;
-        }
+        $this->urls = is_array($url)
+            ? array_merge($this->urls, $url)
+            : array_merge($this->urls, [$url]);
     }
 
     /**
-     * Parse and prepare URLs for indexing.
+     * Parses and validates an array of URLs. Ensures all URLs are absolute by appending the base application URL to relative ones.
+     * Processes each URL to confirm validity and stores the result. Throws exceptions for invalid or excessive URLs.
      *
-     * @throws ExcessUrlsException | NonAbsoluteUrlException
+     * @return void
+     * @throws ExcessUrlsException If the number of URLs exceeds the allowed limit.
+     * @throws NonAbsoluteUrlException If a URL is determined to be invalid or non-absolute after processing.
      */
     private function parseUrls(): void
     {
         $urlApp = Str::of(Config::get('app.url'))
-            ->replaceMatches('#/*$#', '');
+            ->replaceMatches('#/*$#', ''); // Ensure base URL is clean
 
         if (count($this->urls) > 10000) {
             throw new ExcessUrlsException();
         }
 
         foreach ($this->urls as $index => $url) {
-            // If the url does not have a host it is considered relative.
-            // Try to make it absolute by concatenating it with the base url (domain) of the application.
+            // If the URL does not have a host, it is considered relative.
+            // Try to make it absolute by concatenating it with the base URL.
             if (is_null(parse_url($url, PHP_URL_HOST))) {
                 $url = Str::of($urlApp)->append($url);
+            } else {
+                $url = Str::of($url); // Ensure `$url` is always Stringable
             }
 
-            // Check if the url is absolute
+            // Check if the URL is valid and absolute
             if (is_null(parse_url($url, PHP_URL_HOST))) {
                 throw new NonAbsoluteUrlException();
             }
 
-            $this->urls[$index] = $url->toString();
+            $this->urls[$index] = (string) $url; // Convert back to string for final assignment
         }
     }
 
     /**
-     * Get the host for the URLs.
+     * Constructs and returns the request data as an associative array.
+     *
+     * @return array The constructed request data containing key, host, urlList, and optionally keyLocation.
+     */
+    private function buildRequestData(): array
+    {
+        $data = [
+            'key' => $this->key,
+            'host' => $this->getHost(),
+            'urlList' => $this->getUrls(),
+        ];
+
+        if ($this->keyFile) {
+            $data['keyLocation'] = $this->keyFile;
+        }
+
+        return $data;
+    }
+
+    /**
+     * Retrieves the host from the list of URLs or falls back to the application's base URL if no valid host is found.
+     *
+     * @return string The extracted host from the URLs or the application's base URL host. Returns an empty string if the host cannot be determined.
      */
     private function getHost(): string
     {
-        if (isset($this->getUrls()[0])) {
-            $host = parse_url(rawurldecode($this->getUrls()[0]), PHP_URL_HOST);
-        }
+        $urls = $this->getUrls();
+        $host = isset($urls[0]) ? parse_url(rawurldecode($urls[0]), PHP_URL_HOST) : null;
 
-        if (! isset($host)) {
-            $urlApp = Str::of(Config::get('app.urls'))
-                ->replaceMatches('#/*$#', '');
-
-            $host = parse_url($urlApp, PHP_URL_HOST);
+        if (!$host) {
+            $appBaseUrl = Str::of(Config::get('app.url'))->replaceMatches('#/*$#', '');
+            $host = parse_url((string)$appBaseUrl, PHP_URL_HOST);
         }
 
         return $host ?? '';
     }
 
     /**
-     * Process the indexing of URLs.
+     * Processes URL submission to the IndexNow API based on the provided configuration and environment settings.
      *
-     * @return string|array
+     * This method validates the environment, checks spam prevention mechanisms, prepares the URLs,
+     * and sends them to the appropriate search engine endpoint for indexing. Responses are handled
+     * accordingly, and potential exceptions are captured and rethrown as MixedException.
      *
-     * @throws MixedException
+     * @return array|string Returns an array containing the search engine name, response status, info,
+     * and URLs if the submission is successful, or a string message for specific cases like a disabled
+     * environment, a temporary block due to spam prevention, or when no URLs are provided for indexing.
+     * @throws MixedException If an error occurs during the HTTP request to the IndexNow API.
      */
-    private function process()
+    private function process(): array|string
     {
-        if (Config::get('indexnow.ignore_production_environment') !== true && App::isProduction() == false) {
-            return 'Sending requests to IndexNow is currently disabled in the local environment. To enable request sending in any environment, set the property "ignore_production_environment" to true in the configuration file. If you do not have the package configuration file, you can install it using the following command: php artisan vendor:publish --tag="indexnow"';
+        if (! Config::get('indexnow.enable_submissions')) {
+            return 'Enable Submissions is set to false. To allow submissions, please check the configuration file and set it to true.';
         }
 
-        if (PreventSpan::isAllowed() == false) {
+        if (!PreventSpan::isAllowed()) {
             return 'The use of IndexNow has been temporarily blocked to prevent potential spam.';
         }
 
         $this->parseUrls();
+        $urls = $this->getUrls();
 
         $endpoint = Str::of('https://<searchengine>/indexnow')
             ->replace('<searchengine>', $this->getSearchEngine());
-        $response = null;
 
         try {
-            if (count($this->getUrls()) > 1) {
-                $data = [];
-                $data['host'] = $this->getHost();
-                $data['key'] = $this->getKey();
-
-                if ($this->keyFile) {
-                    $data['keyLocation'] = $this->keyFile;
-                }
-
-                $data['urlList'] = $this->getUrls();
-
-                $response = Http::post($endpoint, $data);
-            } elseif (count($this->getUrls()) == 1) {
-                $data = [];
-                $data['url'] = $this->getUrls()[0];
-                $data['key'] = $this->getKey();
-
-                if ($this->keyFile) {
-                    $data['keyLocation'] = $this->keyFile;
-                }
-
-                $endpoint = Str::of($endpoint)
-                    ->append('?')
-                    ->append(http_build_query($data));
-
-                $response = Http::get($endpoint);
+            if (count($urls) > 1) {
+                $response = Http::post((string)$endpoint, $this->buildRequestData());
+            } elseif (count($urls) === 1) {
+                $response = Http::get((string)$endpoint->append('?')->append(http_build_query([
+                    'url' => $urls[0],
+                    'key' => $this->key,
+                    'keyLocation' => $this->keyFile,
+                ])));
             } else {
                 return 'No URLs provided for indexing.';
             }
@@ -254,7 +256,7 @@ class IndexNowService
             'searchengine' => $this->getSearchEngine(),
             'status' => $response->status(),
             'info' => $response->reason(),
-            'urls' => $this->getUrls(),
+            'urls' => $urls,
         ];
     }
 }
